refactor: separate providers
