--- apps/web/app/providers.tsx
@@ -1,18 +1,0 @@
-"use client"
-import { QueryClientProvider } from "@tanstack/react-query"
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools"
-import { SessionProvider } from "next-auth/react"
-import { getQueryClient } from "./get-query-client"
-
-export default function Providers({ children }: { children: React.ReactNode }) {
-  const queryClient = getQueryClient()
-
-  return (
-    <SessionProvider>
-      <QueryClientProvider client={queryClient}>
-        {children}
-        <ReactQueryDevtools />
-      </QueryClientProvider>
-    </SessionProvider>
-  )
-}
