+++ apps/web/lib/providers/react-query-provider.tsx
@@ -1,0 +1,19 @@
+"use client"
+import { getQueryClient } from "@/app/get-query-client"
+import { QueryClientProvider } from "@tanstack/react-query"
+import { ReactQueryDevtools } from "@tanstack/react-query-devtools"
+
+export default function ReactQueryProvider({
+  children
+}: {
+  children: React.ReactNode
+}) {
+  const queryClient = getQueryClient()
+
+  return (
+    <QueryClientProvider client={queryClient}>
+      {children}
+      <ReactQueryDevtools />
+    </QueryClientProvider>
+  )
+}
