+++ package.json
@@ -14,7 +14,7 @@
     "turbo": "^2.5.4",
     "typescript": "5.8.2"
   },
-  "packageManager": "pnpm@10.0.0",
+  "packageManager": "pnpm@10.12.4",
   "engines": {
     "node": ">=18"
   }
