+++ packages/api-client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "api-client",
+  "name": "@repo/api-client",
   "version": "1.0.0",
   "private": true,
   "main": "./src/index.ts",
